TokenGate/pi/requirements.txt
Pillow
smbus2
luma.oled
python-dotenv
web3